+++ examples/data-sources/ceph_cluster_fsid/data-source.tf
@@ -1,0 +1,5 @@
+data "ceph_cluster_fsid" "this" {}
+
+output "ceph_cluster_fsid" {
+  value = data.ceph_cluster_fsid.this.fsid
+}