+++ examples/data-sources/ceph_example/data-source.tf
@@ -1,3 +1,0 @@
-data "ceph_example" "example" {
-  configurable_attribute = "some-value"
-}