+++ examples/data-sources/ceph_health/data-source.tf
@@ -1,0 +1,5 @@
+data "ceph_health" "this" {}
+
+output "ceph_health_status" {
+  value = data.ceph_health.this.health_status
+}